+++ src/lib/promise.ts
