+++ src/lib/shell.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from "node:test";
 import assert from "node:assert/strict";
-import { shellSplit } from "./shell-split.ts";
+import { shellSplit, shellEscape } from "./shell.ts";
 
 describe("shellSplit", () => {
   it("splits a simple string into tokens", () => {
@@ -35,3 +35,30 @@ describe("shellSplit", () => {
     assert.deepEqual(shellSplit("a\nb\nc"), ["a", "b", "c"]);
   });
 });
+
+describe("shellEscape", () => {
+  it("wraps simple string in single quotes", () => {
+    assert.equal(shellEscape("hello"), "'hello'")
+  })
+
+  it("escapes single quotes", () => {
+    assert.equal(shellEscape("it's"), "'it'\\''s'")
+  })
+
+  it("handles empty string", () => {
+    assert.equal(shellEscape(""), "''")
+  })
+
+  it("wraps strings with spaces", () => {
+    assert.equal(shellEscape("hello world"), "'hello world'")
+  })
+
+  it("wraps strings with special shell characters", () => {
+    assert.equal(shellEscape("foo;bar"), "'foo;bar'")
+    assert.equal(shellEscape("$(cmd)"), "'$(cmd)'")
+  })
+
+  it("handles multiple single quotes", () => {
+    assert.equal(shellEscape("a'b'c"), "'a'\\''b'\\''c'")
+  })
+})
