+++ src/lib/shell.ts
@@ -4,3 +4,6 @@
  */
 export const shellSplit = (raw: string): string[] =>
   (raw ?? "").trim().split(/\s+/).filter(Boolean);
+
+export const shellEscape = (value: string): string =>
+  "'" + value.replace(/'/g, "'\\''") + "'"
