--- src/shell.ts
@@ -1,2 +1,0 @@
-export const shellEscape = (value: string): string =>
-  "'" + value.replace(/'/g, "'\\''") + "'"
